requirements.txt
Flask==2.0.1
Flask-BabelEx==0.9.4
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-Security==3.0.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.15.1
Werkzeug==2.0.1
python-dotenv==0.19.1
APScheduler==3.8.1
python-dateutil==2.8.1
cryptography==35.0.0
